Finalize hero section
Added hovering animation to navigation links, sign up button and intro
image. Added appropriate styling for color and font-size to each element
in while ensuring they are all spaced appropriatly on the page. Ensured
that the theme of efficiency in terms of styling stays consistent
through out the beautification of the hero section.