Beautify transportation section
Layed out modes of transportation in one row, provided adequate spacing
between modes and between each mode and it's description. Additionally
added space between the bottom of the section and the page footer in
order to have it display neatly.

In addition to all changes, it was ensured that styling was made
efficient to provid easy readability.